Extract logic into separate files
