+++ package.json
@@ -5,6 +5,7 @@
   "private": true,
   "scripts": {
     "build": "webpack",
+    "watch": "webpack --watch",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Alexander Fuks",
